storage: dict: added dictionary backends
Implemented parts of the dictionary backend which are related to
the expanding (i.e. resizing) of the dictionary. Currently,
dictionary's are resized when a ratio of 1:1 between the amount
of elements and buckets is archieved AND it is allowed to resize OR
the ratio as risen above the FORCE_RESIZE ratio.

Signed-off-by: Michel Megens <dev@michelmegens.net>