storage: dict: implmented inserting to a dictionary
New elements can be added to a dictionary using the dict_add
function. This function searches, or creates an available or new
slot in the dictionary. Elements can be one of four types:

	* pointer
	* unsigned 64-bit integer
	* signed 64-bit integer
	* double (floating point)

This function will automatically resize the dictionary if no
slot is available to store the new element.

Signed-off-by: Michel Megens <dev@michelmegens.net>