storage: dict: implemented element deletion
Dictionary elements can be deleted using the API function
'dict_delete'. Elements will only be deleted after the keys
(the given one and the key found on the index indictated by the
hash) are checked using strcmp. This is the only way to ensure the
right element is deleted. Once the correct element is found it will
be removed and all memory associated with it will be free'd.

Signed-off-by: Michel Megens <dev@michelmegens.net>