storage: dict: redesigned some API calls
Refactored the dict_lookup call. This function now works in a similar
fashion as dict_delete, the user passes a `union entry_data' in the
parameters to store the entry data in. The return value is an error
code. Only trust the data in the entry_data if the return code is
-XFIRE_OK.

storage: database: implemented database_lookup.

Signed-off-by: Michel Megens <dev@michelmegens.net>