project: moved include files to the xfiredb directory
All include headers have been moved to the include/xfiredb directory
(old: include/xfire).

ext: object marking

When objects are no longer required, they will be marked and can
therefore be safely free'd. If the GC try's to free an object that
is still stored in the database, the object will be rebuild from the
data available in the database.

Signed-off-by: Michel Megens <dev@michelmegens.net>