server: workerpool: added reply length indicators
Before the actual reply is sent to the client, the server will inform
the client how long each reply is. For example

15 90 22 105

means that there are 4 reply's with the lengths of 15, 90, 22 and 105
bytes in lenth for each reply. In other words, the client should do 5
reads in total; 1 to get the number of messages and its lengths, and 4
more to get the actual messages.

Signed-off-by: Michel Megens <dev@michelmegens.net>