cmake: fixed a build issue by removing cli.h from CMakeLists
After the removal of cli.h, the build system broke due to it failing
to find that specific header. Fixed by removing the from the search
list.

clients: cli: added stdin authentication

One can now authenticate with the CLI using the `-a' or `--auth' option.
This way, the username and password will be asked through the stdin, and
will not end up in files such as `~/.bash_history'.

Signed-off-by: Michel Megens <dev@michelmegens.net>