clients: cli: fixed some reply processing issue's
Some of the reply's weren't processed properly. Namely when an error
occured on the server side. This has been fixed by adding a numeral
string length to the exception handlers. Also, fgets output is now
stripped of its newline character before its being sent of to the
XFireDB client library.

Signed-off-by: Michel Megens <dev@michelmegens.net>